pyproject.toml
[project]
name = "01-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "chromadb>=1.3.4",
   "google-genai>=1.49.0",
   "markitdown[all]>=0.1.3",
   "openai>=1.109.0",
   "python-dotenv>=1.1.1",
   "sentence-transformers>=5.1.1",
   "streamlit>=1.50.0",
]
